Making request with mixed case headers can mess up request
I.e. when making a request with "Content-Type": "application/multipart",
the content type will not be matched correctly and the serializer for
"application/json" will be incorrectly applied to the body.

The same problem happens for the "accept" header, if the content type
is parsed correctly, but "Accept" in the params uses an uppercase "A".

The workaround for those issues for now is to simply use only lowercase
headers, and no atoms.